Log purged messages, Fix macro usage in DMs & Fix reading empty files (#88)
* Log purged messages

* Fix CommandInvokeError for macros in DMs

* Fix decoding empty files and simplify read json logic

* Apply black formatting

+++ robocop_ng/cogs/macro.py
@@ -22,7 +22,8 @@ class Macro(Cog):
     @commands.cooldown(3, 30, BucketType.user)
     @commands.command(aliases=["m"])
     async def macro(self, ctx: Context, key: str, targets: Greedy[discord.User] = None):
-        await ctx.message.delete()
+        if ctx.guild:
+            await ctx.message.delete()
         if len(key) > 0:
             text = get_macro(self.bot, key)
             if text is not None:
